Refactor: colocate tests, remove smartcard dep, delete example
- Move test files from tests/ to src/ alongside source files
- Update imports to use relative paths
- Remove separate tests/tsconfig.json
- Move smartcard from dependencies to optional peerDependencies
  (library only defines interfaces, doesn't import smartcard)
- Delete examples/demo.ts (README has cleaner example)
- Simplify build scripts (single tsc build)
- Update CI workflow, eslint config, and ignore files

Closes #37
