+++ .github/workflows/ci.yml
@@ -31,8 +31,5 @@ jobs:
             - name: Lint
               run: npm run lint
 
-            - name: Build
-              run: npm run build
-
             - name: Test
               run: npm test
