+++ .npmignore
@@ -1,1 +1,5 @@
 src/
+*.test.js
+*.test.d.ts
+*.test.d.ts.map
+*.test.js.map
