+++ eslint.config.js
@@ -27,7 +27,7 @@ export default tseslint.config(
         },
     },
     {
-        files: ['tests/**/*.ts'],
+        files: ['src/**/*.test.ts'],
         rules: {
             '@typescript-eslint/no-floating-promises': 'off',
             '@typescript-eslint/require-await': 'off',
@@ -35,6 +35,6 @@ export default tseslint.config(
         },
     },
     {
-        ignores: ['dist/', 'dist-tests/', 'node_modules/', '*.js', '*.cjs'],
+        ignores: ['dist/', 'node_modules/', '*.js', '*.cjs'],
     }
 );
