+++ tsconfig.json
@@ -28,6 +28,5 @@
         "resolveJsonModule": true,
         "isolatedModules": true
     },
-    "include": ["src/**/*"],
-    "exclude": ["node_modules", "dist", "**/*.test.ts"]
+    "include": ["src/**/*"]
 }
